Fix handling of LV2 plugin bool parameters
An atom:Bool is stored as int32_t, not float (although the nature of falsey
zero made it mostly work anyway).

+++ libs/ardour/lv2_plugin.cc
@@ -2848,7 +2848,7 @@ LV2Plugin::handle_property_change (const uint32_t key, const LV2_Atom* value)
 {
 	if (_property_values.find (key) != _property_values.end ()) {
 		if (value->type == _uri_map.urids.atom_Bool) {
-			const float* val      = (const float*)LV2_ATOM_BODY_CONST (value);
+			const int32_t* val    = (const int32_t*)LV2_ATOM_BODY_CONST (value);
 			_property_values[key] = Variant (static_cast<bool> (*val));
 		} else if (value->type == _uri_map.urids.atom_Float) {
 			const float* val      = (const float*)LV2_ATOM_BODY_CONST (value);
